Update Square.xml
_(I haven't used IntelliJ for that long and am far from an expert; apologies in advance if I'm missing something obvious or any part of this doesn't make sense!)_

Hi, just wanted to suggest adding the `INSERT_INNER_CLASS_IMPORTS` option here. This maps to the "Insert imports for inner classes" checkbox IntelliJ setting in Editor > Code Style > Java > Imports tab.

With this enabled, and assuming a class like this:
```public class Foo {
  public static class Bar { }
}```
you can write `Ba` and press Option+Space to autocomplete and auto-import `Bar`.

With the flag disabled, you'll end up with `Foo.Bar`.

The outer class qualifier does reduce possible ambiguity, but I'd argue that the majority of the time, you won't need to refer to multiple `Bar` types within a single file. But even when you do, it's simple to achieve the desired `Foo.Bar` outcome by autocompleting both fragments (e.g. type `Fo` -> Option+Space to import `Foo` -> type `.Ba` -> Option+Space to import `Bar`).

On the other hand, with this option disabled, there's no equally-simple way (at least as far as I've found) to automatically perform the static import and enable unqualified references to `Bar`. I've sometimes been able to do this with the IntelliJ lightbulb selector, but other times I've had to jump to the top of the file and add a static import manually, which is quite disruptive. So, since *enabling* this option makes it easy to write code in both styles (either referring to `outerClass.innerClass` or to the unqualified `innerClass` directly), while *disabling* this option makes the latter option difficult, I think enabling this option is a better default.